Do not fixnumify before uploading spec.
Actually this is quite harmful as a string which is actually a number
might be fixnumified where API expects a string. For example,
environment variable specification in POD spec:

env: [{
  name: FOO
  value: 300
}]

What is more, I don't see much a point in this process. Let rely on
users to specify correct types rather than being lazy.